services/fastapi/requirements.txt
fastapi
uvicorn
psycopg2-binary
fastapi
uvicorn
psycopg2-binary
reportlab
